fix block_query_test
